refactor: remove post-mortem analysis, off the value proposition
The leaderboard's M key generated per-model "coach notes" via an analyzer
model: unvalidated LLM prose feeding nothing in the rating or its trust
story, the only untested LLM-call path in the repo, and the only way the
TUI could spend tokens. Deleting it kills the analyzer_model config key,
the analysis.jsonl artifact, and one prompt+schema to maintain; now only
`run` and `rejudge` ever hit the API.

The unfiltered record loader the battle browser shared with the deleted
module moves to data/schemas.py next to BattleRecord (rejudge's loader is
not a substitute: it drops voided rounds, the browser must show them).
Leaderboard and battle-browser screenshots regenerated without the M hint.
